+++ packages/core/src/__tests__/test-generator.ts
@@ -1,8 +1,9 @@
-import { describe, test } from 'vitest';
+import { parseJsx } from '@/parsers/jsx';
+import { Target } from '@/types/config';
+import { BaseTranspilerOptions, TranspilerGenerator } from '@/types/transpiler';
+import { describe, expect, test } from 'vitest';
 import { MitosisComponent, createTypescriptProject, parseSvelte } from '..';
-import { parseJsx } from '../parsers/jsx';
-import { Target } from '../types/config';
-import { BaseTranspilerOptions, TranspilerGenerator } from '../types/transpiler';
+
 const getRawFile = async (filePath: string) => {
   const code = await import(`${filePath}?raw`).then((x) => x.default as string);
   return { code, filePath: ['src', '__tests__', filePath].join('/') };
@@ -49,6 +50,8 @@ const propsType = getRawFile('./data/types/component-props-type.raw.tsx');
 const propsInterface = getRawFile('./data/types/component-props-interface.raw.tsx');
 const preserveTyping = getRawFile('./data/types/preserve-typing.raw.tsx');
 const typeDependency = getRawFile('./data/types/type-dependency.raw.tsx');
+const typeExternalStore = getRawFile('./data/types/type-external-store.raw.tsx');
+const typeGetterStore = getRawFile('./data/types/type-getter-store.raw.tsx');
 
 const defaultProps = getRawFile('./data/default-props/default-props.raw.tsx');
 const defaultPropsOutsideComponent = getRawFile(
@@ -83,6 +86,7 @@ const inputParentBlock = getRawFile('./data/blocks/input-parent.raw.tsx');
 const multipleOnUpdate = getRawFile('./data/blocks/multiple-onUpdate.raw.tsx');
 const multipleOnUpdateWithDeps = getRawFile('./data/blocks/multiple-onUpdateWithDeps.raw.tsx');
 const onInit = getRawFile('./data/blocks/onInit.raw.tsx');
+const onInitPlain = getRawFile('./data/blocks/onInit-plain.raw.tsx');
 const onEvent = getRawFile('./data/blocks/onEvent.raw.tsx');
 const onInitonMount = getRawFile('./data/blocks/onInit-onMount.raw.tsx');
 const onMount = getRawFile('./data/blocks/onMount.raw.tsx');
@@ -191,6 +195,7 @@ const BASIC_TESTS: Tests = {
   Columns: columns,
   onUpdate: onUpdate,
   onInit: onInit,
+  onInitPlain,
   onEvent,
   onUpdateWithDeps: onUpdateWithDeps,
   onMount: onMount,
@@ -201,6 +206,8 @@ const BASIC_TESTS: Tests = {
   defaultPropsOutsideComponent,
   preserveTyping: preserveTyping,
   typeDependency,
+  typeExternalStore,
+  typeGetterStore,
   defaultValsWithTypes: getRawFile('./data/types/component-with-default-values-types.raw.tsx'),
   'import types': builderRenderContent,
   subComponent,
@@ -316,6 +323,7 @@ const IMPORT_TEST: Tests = {
 };
 
 const ANGULAR_TESTS: Tests = {
+  customSelector: getRawFile('./data/angular/custom-selector.raw.tsx'),
   nativeAttributes: getRawFile('./data/angular/native-attributes.raw.tsx'),
   outputEventBinding: getRawFile('./data/angular/output-event-bindings.raw.tsx'),
   dynamicComponent: getRawFile('./data/angular/dynamic-component.raw.tsx'),
@@ -326,6 +334,7 @@ const ANGULAR_TESTS: Tests = {
   stateInit: getRawFile('./data/angular/state-init.raw.tsx'),
   stateInitSequence: getRawFile('./data/angular/state-init-sequence.raw.tsx'),
   useObjectWrapper: getRawFile('./data/angular/use-object-wrapper.raw.tsx'),
+  allSpread: getRawFile('./data/angular/all-spread.raw.tsx'),
 };
 
 const CONTEXT_TEST: Tests = {
@@ -473,13 +482,12 @@ const JSX_TESTS_FOR_TARGET: Partial<Record<Target, Tests[]>> = {
     CONTEXT_TEST,
     BASIC_TESTS,
     SLOTS_TESTS,
-    // ROOT_SHOW_TESTS,
     FORWARD_REF_TESTS,
-    // MULTI_ON_UPDATE_TESTS,
+    MULTI_ON_UPDATE_TESTS,
     FORM_BLOCK_TESTS,
     ADVANCED_REF,
     ON_UPDATE_RETURN,
-    // FOR_SHOW_TESTS
+    FOR_SHOW_TESTS,
   ],
   solid: [
     CONTEXT_TEST,
@@ -580,14 +588,35 @@ export const runTestsForSvelteSyntax = () => {
 
 const tsProject = createTypescriptProject(__dirname + '/tsconfig.json');
 
+const filterTests = (testArray?: Tests[], only?: string[]) =>
+  testArray?.map((tests: Tests) => {
+    if (!only) {
+      return tests;
+    }
+
+    const filteredTests: Tests = {};
+
+    Object.entries(tests).forEach(([key, test]) => {
+      if (only.includes(key)) {
+        filteredTests[key] = test;
+      }
+    });
+
+    return filteredTests;
+  });
+
 export const runTestsForTarget = <X extends BaseTranspilerOptions>({
   target,
   generator,
   options,
+  only,
+  logOutput,
 }: {
   target: Target;
   generator: TranspilerGenerator<X>;
   options: X;
+  logOutput?: boolean;
+  only?: string[]; // Test only some tests based on key
 }) => {
   const configurations: { options: X; testName: string }[] = [
     { options: { ...options, typescript: false }, testName: 'Javascript Test' },
@@ -617,26 +646,28 @@ export const runTestsForTarget = <X extends BaseTranspilerOptions>({
                   typescript: false,
                 },
           ),
-        testsArray: JSX_TESTS_FOR_TARGET[target],
+        testsArray: filterTests(JSX_TESTS_FOR_TARGET[target], only),
       },
       {
         name: 'svelte',
-        parser: async ({ filePath, code }) => parseSvelte(code),
-        testsArray: [SVELTE_SYNTAX_TESTS],
+        parser: async ({ code }) => parseSvelte(code),
+        testsArray: filterTests([SVELTE_SYNTAX_TESTS], only),
       },
     ];
     for (const { name, parser, testsArray } of parsers) {
       if (testsArray) {
         describe(name, () => {
-          if (name === 'jsx' && options.typescript === false) {
-            test('Remove Internal mitosis package', async () => {
-              const t = await basicMitosis;
-              const component = parseJsx(t.code, {
-                compileAwayPackages: ['@dummy/custom-mitosis'],
+          if (!only) {
+            if (name === 'jsx' && options.typescript === false) {
+              test('Remove Internal mitosis package', async () => {
+                const t = await basicMitosis;
+                const component = parseJsx(t.code, {
+                  compileAwayPackages: ['@dummy/custom-mitosis'],
+                });
+                const output = generator(options)({ component, path: t.filePath });
+                expect(output).toMatchSnapshot();
               });
-              const output = generator(options)({ component, path: t.filePath });
-              expect(output).toMatchSnapshot();
-            });
+            }
           }
           describe(testName, () => {
             testsArray.forEach((tests) => {
@@ -654,7 +685,13 @@ export const runTestsForTarget = <X extends BaseTranspilerOptions>({
                     expect(getOutput).toThrowError();
                   } else {
                     try {
-                      expect(getOutput()).toMatchSnapshot();
+                      const output = getOutput();
+                      if (logOutput) {
+                        process.stdout.write(`--- Start: ${key} ---\n\n`);
+                        process.stdout.write(output);
+                        process.stdout.write(`--- End: ${key} ---\n\n`);
+                      }
+                      expect(output).toMatchSnapshot();
                     } catch (error) {
                       expect(getOutput).toThrowErrorMatchingSnapshot();
                     }