+++ packages/core/src/generators/angular/index.ts
@@ -1,6 +1,7 @@
 import { SELF_CLOSING_HTML_TAGS, VALID_HTML_TAGS } from '@/constants/html_tags';
 import { dedent } from '@/helpers/dedent';
 import { fastClone } from '@/helpers/fast-clone';
+import { getChildComponents } from '@/helpers/get-child-components';
 import { getComponentsUsed } from '@/helpers/get-components-used';
 import { getCustomImports } from '@/helpers/get-custom-imports';
 import { getPropFunctions } from '@/helpers/get-prop-functions';
@@ -8,6 +9,7 @@ import { getProps } from '@/helpers/get-props';
 import { getPropsRef } from '@/helpers/get-props-ref';
 import { getRefs } from '@/helpers/get-refs';
 import { getStateObjectStringFromComponent } from '@/helpers/get-state-object-string';
+import { getTypedFunction } from '@/helpers/get-typed-function';
 import { indent } from '@/helpers/indent';
 import { isMitosisNode } from '@/helpers/is-mitosis-node';
 import { isUpperCase } from '@/helpers/is-upper-case';
@@ -33,7 +35,7 @@ import {
   runPreJsonPlugins,
 } from '@/modules/plugins';
 import { hashCodeAsString } from '@/symbols/symbol-processor';
-import { MitosisComponent } from '@/types/mitosis-component';
+import { BaseHook, MitosisComponent } from '@/types/mitosis-component';
 import { Binding, MitosisNode, checkIsForNode } from '@/types/mitosis-node';
 import { TranspilerGenerator } from '@/types/transpiler';
 import { flow, pipe } from 'fp-ts/lib/function';
@@ -42,7 +44,15 @@ import traverse from 'neotraverse/legacy';
 import { format } from 'prettier/standalone';
 import isChildren from '../../helpers/is-children';
 import { stringifySingleScopeOnMount } from '../helpers/on-mount';
-import { HELPER_FUNCTIONS, getAppropriateTemplateFunctionKeys } from './helpers';
+import {
+  HELPER_FUNCTIONS,
+  addCodeToOnInit,
+  addCodeToOnUpdate,
+  getAppropriateTemplateFunctionKeys,
+  getDefaultProps,
+  makeReactiveState,
+  transformState,
+} from './helpers';
 import {
   AngularBlockOptions,
   BUILT_IN_COMPONENTS,
@@ -51,16 +61,23 @@ import {
 } from './types';
 
 const mappers: {
-  [key: string]: (root: MitosisComponent, json: MitosisNode, options: ToAngularOptions) => string;
+  [key: string]: (
+    root: MitosisComponent,
+    json: MitosisNode,
+    options: ToAngularOptions,
+    blockOptions: AngularBlockOptions,
+  ) => string;
 } = {
-  Fragment: (root, json, options) => {
+  Fragment: (root, json, options, blockOptions) => {
     return `<ng-container>${json.children
-      .map((item) => blockToAngular({ root, json: item, options }))
+      .map((item) => blockToAngular({ root, json: item, options, blockOptions }))
       .join('\n')}</ng-container>`;
   },
-  Slot: (root, json, options) => {
+  Slot: (root, json, options, blockOptions) => {
     const renderChildren = () =>
-      json.children?.map((item) => blockToAngular({ root, json: item, options })).join('\n');
+      json.children
+        ?.map((item) => blockToAngular({ root, json: item, options, blockOptions }))
+        .join('\n');
 
     return `<ng-content ${Object.entries({ ...json.bindings, ...json.properties })
       .map(([binding, value]) => {
@@ -194,26 +211,28 @@ const processEventBinding = (key: string, code: string, nodeName: string, custom
 const stringifyBinding =
   (node: MitosisNode, options: ToAngularOptions, blockOptions: AngularBlockOptions) =>
   ([key, binding]: [string, Binding | undefined]) => {
-    if (options.state === 'inline-with-wrappers' && binding?.type === 'spread') {
-      return;
-    }
     if (key.startsWith('$') || key.startsWith('"') || key === 'key') {
       return;
     }
+    if (binding?.type === 'spread') {
+      return;
+    }
+
     const keyToUse = BINDINGS_MAPPER[key] || key;
     const { code, arguments: cusArgs = ['event'] } = binding!;
     // TODO: proper babel transform to replace. Util for this
 
     if (keyToUse.startsWith('on')) {
       const { event, value } = processEventBinding(keyToUse, code, node.name, cusArgs[0]);
-      return ` (${event})="${value}"`;
+      // Angular events are all lowerCased
+      return ` (${event.toLowerCase()})="${value}"`;
     } else if (keyToUse === 'class') {
       return ` [class]="${code}" `;
-    } else if (keyToUse === 'ref') {
+    } else if (keyToUse === 'ref' || keyToUse === 'spreadRef') {
       return ` #${code} `;
     } else if (
       (VALID_HTML_TAGS.includes(node.name.trim()) || keyToUse.includes('-')) &&
-      !blockOptions.nativeAttributes.includes(keyToUse) &&
+      !blockOptions.nativeAttributes?.includes(keyToUse) &&
       !Object.values(BINDINGS_MAPPER).includes(keyToUse)
     ) {
       // standard html elements need the attr to satisfy the compiler in many cases: eg: svg elements and [fill]
@@ -298,7 +317,7 @@ export const blockToAngular = ({
   const childComponents = blockOptions?.childComponents || [];
 
   if (mappers[json.name]) {
-    return mappers[json.name](root, json, options);
+    return mappers[json.name](root, json, options, blockOptions);
   }
 
   if (isChildren({ node: json })) {
@@ -384,10 +403,7 @@ export const blockToAngular = ({
         type: 'property',
       };
       if (!root.hooks.onInit?.code.includes(inputsPropsStateName)) {
-        if (!root.hooks.onInit) {
-          root.hooks.onInit = { code: '' };
-        }
-        root.hooks.onInit.code += `\nthis.${inputsPropsStateName} = {${allProps}};\n`;
+        addCodeToOnInit(root, `this.${inputsPropsStateName} = {${allProps}};`);
       }
       if (
         !root.hooks.onUpdate
@@ -395,12 +411,7 @@ export const blockToAngular = ({
           .join('')
           .includes(inputsPropsStateName)
       ) {
-        if (!root.hooks.onUpdate) {
-          root.hooks.onUpdate = [];
-        }
-        root.hooks.onUpdate.push({
-          code: `this.${inputsPropsStateName} = {${allProps}}`,
-        });
+        addCodeToOnUpdate(root, `this.${inputsPropsStateName} = {${allProps}};`);
       }
       allProps = `${inputsPropsStateName}`;
     } else {
@@ -420,19 +431,72 @@ export const blockToAngular = ({
       : json.name;
     str += `<${elSelector} `;
 
-    // TODO: spread support for angular
-    // if (json.bindings._spread) {
-    //   str += `v-bind="${stripStateAndPropsRefs(
-    //     json.bindings._spread as string,
-    //   )}"`;
-    // }
-
     for (const key in json.properties) {
       if (key.startsWith('$')) {
         continue;
       }
       const value = json.properties[key];
       str += ` ${key}="${value}" `;
+    }
+
+    for (const key in json.bindings) {
+      if (json.bindings[key]?.type === 'spread' && VALID_HTML_TAGS.includes(json.name.trim())) {
+        if (json.bindings[key]?.code === 'this') {
+          // if its an arbitrary { ...props } spread then we skip because Angular needs a named prop to be defined
+          continue;
+        }
+
+        let refName = '';
+        if (json.bindings['spreadRef']?.code) {
+          refName = json.bindings['spreadRef'].code;
+        } else {
+          const spreadRefIndex = root.meta._spreadRefIndex || 0;
+          refName = `elRef${spreadRefIndex}`;
+          root.meta._spreadRefIndex = (spreadRefIndex as number) + 1;
+          json.bindings['spreadRef'] = { code: refName, type: 'single' };
+          root.refs[refName] = { argument: '' };
+        }
+        json.bindings['spreadRef'] = { code: refName, type: 'single' };
+        root.refs[refName] = { argument: '' };
+        root.meta.onViewInit = (root.meta.onViewInit || { code: '' }) as BaseHook;
+        let spreadCode = '';
+        let changesCode = '';
+        if (json.bindings[key]?.code.startsWith('{')) {
+          json.meta._spreadStateRef = json.meta._spreadStateRef || 0;
+          const name = `${refName}_state_${json.meta._spreadStateRef}`;
+          json.meta._spreadStateRef = (json.meta._spreadStateRef as number) + 1;
+          makeReactiveState(root, name, `this.${name} = ${json.bindings[key]?.code};`);
+          spreadCode = `this.${name}`;
+          changesCode = `changes['${spreadCode.replace('this.', '')}']?.currentValue`;
+        } else {
+          spreadCode = `${json.bindings[key]?.code}`;
+          changesCode = `changes['${spreadCode.replace('this.', '')}']?.currentValue`;
+        }
+        if (!root.compileContext) {
+          root.compileContext = {
+            angular: {
+              hooks: {
+                ngAfterViewInit: {
+                  code: '',
+                },
+              },
+            },
+          };
+        }
+        root.compileContext.angular!.hooks!.ngAfterViewInit.code += `\nthis.setAttributes(this.${refName}?.nativeElement, ${spreadCode});`;
+        addCodeToOnUpdate(
+          root,
+          `this.setAttributes(this.${refName}?.nativeElement, ${spreadCode}${
+            changesCode ? `, ${changesCode}` : ''
+          });`,
+        );
+        if (!root.state['setAttributes']) {
+          root.state['setAttributes'] = {
+            code: HELPER_FUNCTIONS(options?.typescript).setAttributes as string,
+            type: 'method',
+          };
+        }
+      }
     }
 
     const stringifiedBindings = Object.entries(json.bindings)
@@ -573,14 +637,11 @@ const handleBindings = (
     } else if (item.bindings[key]?.code) {
       if (item.bindings[key]?.type !== 'spread' && !key.startsWith('on')) {
         json.state[newBindingName] = { code: 'null', type: 'property' };
-        if (!json.hooks['onInit']?.code) {
-          json.hooks['onInit'] = { code: '' };
-        }
-        json.hooks['onInit'].code += `\nstate.${newBindingName} = ${item.bindings[key]!.code};\n`;
-        json.hooks['onUpdate'] = json.hooks['onUpdate'] || [];
-        json.hooks['onUpdate'].push({
-          code: `state.${newBindingName} = ${item.bindings[key]!.code}`,
-        });
+        makeReactiveState(
+          json,
+          newBindingName,
+          `this.${newBindingName} = ${item.bindings[key]!.code}`,
+        );
         item.bindings[key]!.code = `state.${newBindingName}`;
       } else if (key.startsWith('on')) {
         const { arguments: cusArgs = ['event'] } = item.bindings[key]!;
@@ -595,17 +656,11 @@ const handleBindings = (
           item.bindings[key]!.code = `state.${newBindingName}(${cusArgs.join(', ')})`;
         }
       } else {
-        json.state[newBindingName] = { code: `null`, type: 'property' };
-        if (!json.hooks['onInit']?.code) {
-          json.hooks['onInit'] = { code: '' };
-        }
-        json.hooks['onInit'].code += `\nstate.${newBindingName} = {...(${
-          item.bindings[key]!.code
-        })};\n`;
-        json.hooks['onUpdate'] = json.hooks['onUpdate'] || [];
-        json.hooks['onUpdate'].push({
-          code: `state.${newBindingName} = {...(${item.bindings[key]!.code})}`,
-        });
+        makeReactiveState(
+          json,
+          newBindingName,
+          `state.${newBindingName} = {...(${item.bindings[key]!.code})}`,
+        );
         item.bindings[newBindingName] = item.bindings[key];
         item.bindings[key]!.code = `state.${newBindingName}`;
         delete item.bindings[key];
@@ -672,25 +727,6 @@ const classPropertiesPlugin = () => ({
   },
 });
 
-// if any state "property" is trying to access state.* or props.*
-// then we need to move them to onInit where they can be accessed
-const transformState = (json: MitosisComponent) => {
-  Object.entries(json.state)
-    .reverse()
-    .forEach(([key, value]) => {
-      if (value?.type === 'property') {
-        if (value.code && (value.code.includes('state.') || value.code.includes('props.'))) {
-          const code = stripStateAndPropsRefs(value.code, { replaceWith: 'this' });
-          json.state[key]!.code = 'null';
-          if (!json.hooks.onInit?.code) {
-            json.hooks.onInit = { code: '' };
-          }
-          json.hooks.onInit.code = `\nthis.${key} = ${code};\n${json.hooks.onInit.code}`;
-        }
-      }
-    });
-};
-
 export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
   (userOptions = {}) =>
   ({ component: _component }) => {
@@ -714,7 +750,7 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
     });
     options.plugins = [
       ...(options.plugins || []),
-      CODE_PROCESSOR_PLUGIN((codeType) => {
+      CODE_PROCESSOR_PLUGIN((codeType, _, node) => {
         switch (codeType) {
           case 'hooks':
             return flow(
@@ -739,11 +775,18 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
             );
 
           case 'bindings':
-            return (code) => {
+            return (code, key) => {
+              // we create a separate state property for spread binding and use ref to attach the attributes
+              // so we need to use `this.` inside the class to access state and props
+              const isSpreadAttributeBinding =
+                node?.bindings[key]?.type === 'spread' &&
+                VALID_HTML_TAGS.includes(node.name.trim());
+
               const newLocal = processAngularCode({
                 contextVars: [],
                 outputVars,
                 domRefs: [], // the template doesn't need the this keyword.
+                replaceWith: isSpreadAttributeBinding ? 'this' : undefined,
               })(code);
               return newLocal.replace(/"/g, '&quot;');
             };
@@ -767,16 +810,9 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
     }
 
     const [forwardProp, hasPropRef] = getPropsRef(json, true);
-    const childComponents: string[] = [];
     const propsTypeRef = json.propsTypeRef !== 'any' ? json.propsTypeRef : undefined;
 
-    json.imports.forEach(({ imports }) => {
-      Object.keys(imports).forEach((key) => {
-        if (imports[key] === 'default') {
-          childComponents.push(key);
-        }
-      });
-    });
+    const childComponents: string[] = getChildComponents(json);
 
     const customImports = getCustomImports(json);
 
@@ -795,7 +831,6 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
       }
       return `public ${variableName} : ${variableType}`;
     });
-    const hasConstructor = Boolean(injectables.length);
 
     const props = getProps(json);
     // prevent jsx props from showing up as @Input
@@ -878,25 +913,37 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
 
     const dataString = getStateObjectStringFromComponent(json, {
       format: 'class',
-      valueMapper: processAngularCode({
-        replaceWith: 'this',
-        contextVars,
-        outputVars,
-        domRefs: Array.from(domRefs),
-        stateVars,
-      }),
+      withType: options.typescript,
+      valueMapper: (code, type, typeParameter) => {
+        let value = code;
+        if (type !== 'data') {
+          value = getTypedFunction(code, options.typescript, typeParameter);
+        }
+
+        return processAngularCode({
+          replaceWith: 'this',
+          contextVars,
+          outputVars,
+          domRefs: Array.from(domRefs),
+          stateVars,
+        })(value);
+      },
     });
+
+    const refsForObjSpread = getRefs(json, 'spreadRef');
 
     const hostDisplayCss = options.visuallyIgnoreHostElement ? ':host { display: contents; }' : '';
     const styles = css.length ? [hostDisplayCss, css].join('\n') : hostDisplayCss;
 
     // Preparing built in component metadata parameters
     const componentMetadata: Record<string, any> = {
-      selector: `'${kebabCase(json.name || 'my-component')}, ${json.name}'`,
+      selector: useMetadata?.angular?.selector
+        ? `'${useMetadata?.angular?.selector}'`
+        : `'${kebabCase(json.name || 'my-component')}'`,
       template: `\`
-        ${indent(dynamicTemplate, 8).replace(/`/g, '\\`').replace(/\$\{/g, '\\${')}
-        ${indent(template, 8).replace(/`/g, '\\`').replace(/\$\{/g, '\\${')}
-        \``,
+      ${indent(dynamicTemplate, 8).replace(/`/g, '\\`').replace(/\$\{/g, '\\${')}
+      ${indent(template, 8).replace(/`/g, '\\`').replace(/\$\{/g, '\\${')}
+      \``,
       ...(styles
         ? {
             styles: `[\`${indent(styles, 8)}\`]`,
@@ -915,24 +962,17 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
       componentMetadata[key] = value;
     });
 
-    const getPropsDefinition = ({ json }: { json: MitosisComponent }) => {
-      if (!json.defaultProps) return '';
-      const defalutPropsString = Object.keys(json.defaultProps)
-        .map((prop) => {
-          const value = json.defaultProps!.hasOwnProperty(prop)
-            ? json.defaultProps![prop]?.code
-            : 'undefined';
-          return `${prop}: ${value}`;
-        })
-        .join(',');
-      return `const defaultProps = {${defalutPropsString}};\n`;
-    };
+    const hasConstructor =
+      Boolean(injectables.length) || dynamicComponents.size || refsForObjSpread.size;
 
     const angularCoreImports = [
       ...(outputs.length ? ['Output', 'EventEmitter'] : []),
       ...(options?.experimental?.inject ? ['Inject', 'forwardRef'] : []),
       'Component',
-      ...(domRefs.size || dynamicComponents.size ? ['ViewChild', 'ElementRef'] : []),
+      ...(domRefs.size || dynamicComponents.size || refsForObjSpread.size
+        ? ['ViewChild', 'ElementRef']
+        : []),
+      ...(refsForObjSpread.size ? ['Renderer2'] : []),
       ...(props.size ? ['Input'] : []),
       ...(dynamicComponents.size ? ['ViewContainerRef', 'TemplateRef'] : []),
       ...(json.hooks.onUpdate?.length && options.typescript ? ['SimpleChanges'] : []),
@@ -943,7 +983,7 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
     ${options.standalone ? `import { CommonModule } from '@angular/common';` : ''}
 
     ${json.types ? json.types.join('\n') : ''}
-    ${getPropsDefinition({ json })}
+    ${getDefaultProps(json)}
     ${renderPreComponent({
       explicitImportFileExtension: options.explicitImportFileExtension,
       component: json,
@@ -983,6 +1023,13 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
             `@ViewChild('${refName}') ${refName}${options.typescript ? '!: ElementRef' : ''}`,
         )
         .join('\n')}
+
+      ${Array.from(refsForObjSpread)
+        .map(
+          (refName) =>
+            `@ViewChild('${refName}') ${refName}${options.typescript ? '!: ElementRef' : ''}`,
+        )
+        .join('\n')}
       
       ${Array.from(dynamicComponents)
         .map(
@@ -996,6 +1043,11 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
         .join('\n')}
 
       ${dynamicComponents.size ? `myContent${options.typescript ? '?: any[][];' : ''}` : ''}
+      ${
+        refsForObjSpread.size
+          ? `_listenerFns = new Map${options.typescript ? '<string, () => void>' : ''}()`
+          : ''
+      }
 
       ${dataString}
 
@@ -1020,11 +1072,15 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
         .join('\n')}
 
       ${
-        !hasConstructor && !dynamicComponents.size
+        !hasConstructor
           ? ''
           : `constructor(\n${injectables.join(',\n')}${
               dynamicComponents.size
                 ? `\nprivate vcRef${options.typescript ? ': ViewContainerRef' : ''},\n`
+                : ''
+            }${
+              refsForObjSpread.size
+                ? `\nprivate renderer${options.typescript ? ': Renderer2' : ''},\n`
                 : ''
             }) {}
           `
@@ -1067,6 +1123,17 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
       }
 
       ${
+        // hooks specific to Angular
+        json.compileContext?.angular?.hooks
+          ? Object.entries(json.compileContext?.angular?.hooks).map(([key, value]) => {
+              return `${key}() {
+            ${value.code}
+          }`;
+            })
+          : ''
+      }
+
+      ${
         !json.hooks.onUpdate?.length
           ? ''
           : `ngOnChanges(changes${options.typescript ? ': SimpleChanges' : ''}) {
@@ -1081,10 +1148,15 @@ export const componentToAngular: TranspilerGenerator<ToAngularOptions> =
       }
 
       ${
-        !json.hooks.onUnMount
+        !json.hooks.onUnMount && !refsForObjSpread.size
           ? ''
           : `ngOnDestroy() {
-              ${json.hooks.onUnMount.code}
+              ${json.hooks.onUnMount?.code || ''}
+              ${
+                refsForObjSpread.size
+                  ? `for (const fn of this._listenerFns.values()) { fn(); }`
+                  : ''
+              }
             }`
       }