+++ packages/core/src/generators/liquid/types.ts
@@ -1,0 +1,7 @@
+import { BaseTranspilerOptions } from '@/types/transpiler';
+
+export interface ToLiquidOptions extends BaseTranspilerOptions {
+  reactive?: boolean;
+}
+
+export type LiquidMetadata = {};