+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,75 @@
 # @builder.io/mitosis-cli
 
+## 0.5.3
+
+### Patch Changes
+
+- Updated dependencies [82c79fd]
+  - @builder.io/mitosis@0.5.3
+
+## 0.5.2
+
+### Patch Changes
+
+- Updated dependencies [dad8e2f]
+  - @builder.io/mitosis@0.5.2
+
+## 0.5.1
+
+### Patch Changes
+
+- Updated dependencies [dbd5a50]
+  - @builder.io/mitosis@0.5.1
+
+## 0.5.0
+
+### Patch Changes
+
+- Updated dependencies [7e2c95f]
+- Updated dependencies [4171a19]
+- Updated dependencies [d59d328]
+  - @builder.io/mitosis@0.5.0
+
+## 0.4.7
+
+### Patch Changes
+
+- Updated dependencies [068efab]
+  - @builder.io/mitosis@0.4.7
+
+## 0.4.6
+
+### Patch Changes
+
+- Updated dependencies [a7fd87f]
+- Updated dependencies [a7fd87f]
+  - @builder.io/mitosis@0.4.6
+
+## 0.4.5
+
+### Patch Changes
+
+- Updated dependencies [428b3ab]
+  - @builder.io/mitosis@0.4.5
+
+## 0.4.4
+
+### Patch Changes
+
+- Updated dependencies [ad7e576]
+- Updated dependencies [52dc749]
+  - @builder.io/mitosis@0.4.4
+
+## 0.4.3
+
+### Patch Changes
+
+- Updated dependencies [1bf28ea]
+- Updated dependencies [814d171]
+- Updated dependencies [531c15c]
+- Updated dependencies [84038d5]
+  - @builder.io/mitosis@0.4.3
+
 ## 0.4.2
 
 ### Patch Changes